Consolidate kingdom scripts
