+++ audit_log.html
@@ -41,7 +41,7 @@ Developer: Deathsgift66
     // Developer: Deathsgift66
     // Hardened Admin Audit Log Page — with Supabase auth, loading, error handling, and formatting
     import { escapeHTML, authJsonFetch } from '/Javascript/utils.js';
-    import { applyKingdomLinks } from '/Javascript/kingdom_name_linkify.js';
+    import { applyKingdomLinks } from '/Javascript/kingdom.js';
     import { getEnvVar } from '/Javascript/env.js';
 
     import { supabase } from '/Javascript/supabaseClient.js';
