+++ edit_kingdom.html
@@ -138,7 +138,7 @@ Developer: Deathsgift66
   <div id="toast" class="toast-notification" role="status" aria-live="polite"></div>
 
   <!-- Page-specific script -->
-  <script type="module" src="/Javascript/edit_kingdom.js"></script>
+  <script type="module" src="/Javascript/kingdom.js"></script>
 
   <!-- Backend route definition for reference -->
   <script type="text/python">
