+++ message.html
@@ -42,7 +42,7 @@ Developer: Deathsgift66
 // Inbox Viewer + Controls
 
 import { supabase } from '/Javascript/supabaseClient.js';
-import { applyKingdomLinks } from '/Javascript/kingdom_name_linkify.js';
+import { applyKingdomLinks } from '/Javascript/kingdom.js';
 import { escapeHTML, sanitizeHTML } from '/Javascript/utils.js';
 
 const list = document.getElementById('message-list');
