+++ projects.html
@@ -49,7 +49,7 @@ Developer: Deathsgift66
   <script src="/Javascript/apiHelper.js" type="module"></script>
   <script src="/Javascript/navLoader.js" type="module"></script>
   <script src="/Javascript/resourceBar.js" type="module"></script>
-  <script src="/Javascript/projects_kingdom.js" type="module"></script>
+  <script src="/Javascript/kingdom.js" type="module"></script>
 </head>
 
 <body>
