+++ spy_log.html
@@ -37,7 +37,7 @@ Developer: Deathsgift66
 
     import { supabase } from '/Javascript/supabaseClient.js';
     import { escapeHTML, showToast } from '/Javascript/utils.js';
-    import { applyKingdomLinks } from '/Javascript/kingdom_name_linkify.js';
+    import { applyKingdomLinks } from '/Javascript/kingdom.js';
 
     let realtimeChannel = null;
     let currentSession = null;
